Added doc section about security hash
md5sum is kind of weak, but given that usually, AFAIK, the md5 is
cracked appending data, it would overflow the parameters very quickly
and I guess it could be considered secure.

Also, finding a sensible clashing hash requires too much work, and we are going
to check the amount on success anyway.