Refactoring.
Moved the transaction details in a dedicated object.

Given that the Ipayment parent is precious, because it must compile
the wsdl, and will be reused, we don't want to have stale data around.

So, the main object will keep only the fixed values, ro, creating
a new Transaction object every time anew.

In this way it will be more extensible without messing too much